Moved town.js to typescript
